program/requirements.txt
pretrainedmodels==0.7.4
pytorch_gradcam==0.2.1
torch==1.3.0
numpy==1.17.2
requests==2.22.0
torchvision==0.4.1
pydicom==1.4.2
Pillow==6.2.2
